MNMS-529 fix: 서버 합치도록 수정(오류 있음)
