+++ userchatbot_ai/chatbot_app.py
@@ -3,8 +3,10 @@ import sys
 import json
 import requests
 from dotenv import load_dotenv
+from userchatbot_ai.config import invoke_url, headers
 from langchain_aws import BedrockEmbeddings
-from langchain_chroma import Chroma
+# from langchain_chroma import Chroma
+from langchain.vectorstores import Chroma
 
 # --- 1. 경로 및 환경 설정 ---
 current_path = os.path.dirname(os.path.abspath(__file__))
@@ -17,11 +19,11 @@ from userchatbot_ai.config import headers
 dotenv_path = os.path.join(project_root, '.env')
 load_dotenv(dotenv_path=dotenv_path)
 
-# Claude 3 Haiku 모델 엔드포인트 (Bedrock)
-invoke_url = (
-    "https://bedrock-runtime.us-east-1.amazonaws.com/"
-    "model/anthropic.claude-3-haiku-20240307-v1:0/invoke"
-)
+# # Claude 3 Haiku 모델 엔드포인트 (Bedrock)
+# invoke_url = (
+#     "https://bedrock-runtime.us-east-1.amazonaws.com/"
+#     "model/anthropic.claude-3-haiku-20240307-v1:0/invoke"
+# )
 
 # --- 2. 임베딩 및 ChromaDB 로드 ---
 embeddings = BedrockEmbeddings(model_id="amazon.titan-embed-text-v2:0")
